% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_objects.R
\name{build_model_objects}
\alias{build_model_objects}
\title{Build list of model objects}
\usage{
build_model_objects(formulas, data, model = stats::lm, ...)
}
\arguments{
\item{formulas}{List of model formulas}

\item{data}{Dataframe with variables corresponding to the components of `formulas`}

\item{model}{Function that generates model objects from formulas}

\item{...}{args to be passed to `model`}
}
\value{
A list of model objects
}
\description{
Build list of model objects
}
