% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{check_models}
\alias{check_models}
\title{Run code and intercept errors due to invalid inputs}
\usage{
check_models(code)
}
\arguments{
\item{code}{Function to be checked}
}
\value{
The output of code, if ran successfully
}
\description{
Run code and intercept errors due to invalid inputs
}
