% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_list.R
\name{print_list}
\alias{print_list}
\title{Print a named list using markdown-style header tags}
\usage{
print_list(list_to_print, level = 4)
}
\arguments{
\item{level}{Markdown header level}

\item{X}{List of objects to print}
}
\description{
Print a named list using markdown-style header tags
}
